requirements.txt
viam-sdk
torch
transformers
pillow